+++ ostap/core/config.py
@@ -55,6 +55,7 @@ config [ 'Canvas'   ] = { 'Width'       :  '1000' , 'Height'       :  '800' ,
 config [ 'Fit Draw' ] = {}
 config [ 'Parallel' ] = {}
 
+
 ## the list of processes config files 
 files_read = config.read (
     _config.config_files +