+++ ostap/io/lzshelve.py
@@ -146,13 +146,9 @@ __all__ = (
 from sys import version_info as python_version 
 # =============================================================================
 try:
-    from cPickle   import Pickler, Unpickler, HIGHEST_PROTOCOL
+    from cPickle   import Pickler, Unpickler
 except ImportError:
-    from  pickle   import Pickler, Unpickler, HIGHEST_PROTOCOL 
-# =============================================================================
-## to be compatible between  Python2 and Python3 
-PROTOCOL = 2
-ENCODING = 'utf-8'
+    from  pickle   import Pickler, Unpickler
 # =============================================================================
 if python_version.major > 2  :
     from io import BytesIO
@@ -162,11 +158,9 @@ else :
     except ImportError:
         from  StringIO import StringIO as BytesIO    
 # ==============================================================================
-import os, sys
-import lzma        ## use lzma to compress DB-content 
-import shelve      ## 
-import shutil
-from   ostap.io.compress_shelve import CompressShelf
+import os, sys, shelve, shutil 
+import lzma  ## use lzma to compress DB-content 
+from   ostap.io.compress_shelve import CompressShelf, ENCODING, PROTOCOL, HIGHEST_PROTOCOL
 from   ostap.io.dbase           import TmpDB 
 # =============================================================================
 from ostap.logger.logger import getLogger